Switch positional_error type to float to match server

+++ net/client/client.py
@@ -302,7 +302,7 @@ if __name__ == '__main__':
     parser.add_option('--dec', dest='center_dec', type=float, help='Dec center')
     parser.add_option('--radius', dest='radius', type=float, help='Search radius around RA,Dec center')
     parser.add_option('--downsample', dest='downsample_factor', type=int, help='Downsample image by this factor')
-    parser.add_option('--positional_error', dest='positional_error', type=int, help='How many pixels a star may be from where it should be.')
+    parser.add_option('--positional_error', dest='positional_error', type=float, help='How many pixels a star may be from where it should be.')
     parser.add_option('--parity', dest='parity', choices=('0','1'), help='Parity (flip) of image')
     parser.add_option('--tweak-order', dest='tweak_order', type=int, help='SIP distortion order (default: 2)')
     parser.add_option('--crpix-center', dest='crpix_center', action='store_true', default=None, help='Set reference point to center of image?')
